+++ jord.cabal
@@ -2,7 +2,7 @@
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 3b9fe19cfce409ace822fec329109a5d9dd9460b7e8f8bcda2c726e532e4dbae
+-- hash: b8c8ecb9e39c0fa280b26df93858f7feef4a8190a643277b92c3efb2d4aaf49c
 
 name:           jord
 version:        0.3.0.0
@@ -38,10 +38,8 @@ library
       Data.Geo.Jord.Geodetics
       Data.Geo.Jord.LatLong
       Data.Geo.Jord.Length
-      Data.Geo.Jord.NedVector
       Data.Geo.Jord.NVector
       Data.Geo.Jord.Quantity
-      Data.Geo.Jord.SGeodetics
       Data.Geo.Jord.Transform
       Data.Geo.Jord.Vector3d
   other-modules:
@@ -73,11 +71,10 @@ test-suite jord-test
       Data.Geo.Jord.AngleSpec
       Data.Geo.Jord.EarthSpec
       Data.Geo.Jord.EvalSpec
+      Data.Geo.Jord.FramesSpec
       Data.Geo.Jord.GeodeticsSpec
       Data.Geo.Jord.LatLongSpec
       Data.Geo.Jord.LengthSpec
-      Data.Geo.Jord.NedVectorSpec
-      Data.Geo.Jord.SGeodeticsSpec
       Data.Geo.Jord.TransformSpec
       Paths_jord
   hs-source-dirs:
